+++ examples/room_shoebox_3d.py
@@ -15,7 +15,7 @@ room_dim = [5, 4, 6]
 # Create the shoebox
 shoebox = pra.ShoeBox(
     room_dim,
-    absorption=0.2,
+    reflection=0.8,
     fs=16000,
     max_order=15,
     )